# agent-java/src/main/java/com/statedriven/agent/AgentController.java
package com.statedriven.agent;
import com.fasterxml.jackson.databind.*;import com.fasterxml.jackson.databind.node.*;import java.util.*;import java.util.concurrent.ConcurrentHashMap;import org.springframework.http.*;import org.springframework.http.codec.ServerSentEvent;import org.springframework.web.bind.annotation.*;import org.springframework.web.server.ResponseStatusException;import reactor.core.publisher.*;
@RestController final class AgentController {
 private final ObjectMapper json;private final ConfigRepository configs;private final SessionRepository sessions;private final AgentEngine engine;private final Map<String,Object> locks=new ConcurrentHashMap<>();AgentController(ObjectMapper j,ConfigRepository c,SessionRepository s,AgentEngine e){json=j;configs=c;sessions=s;engine=e;}
 @GetMapping("/health") Map<String,Object> health(){return Map.of("status","ok","config_root",configs.root().toString(),"available_configs",configs.list().size());}
 @GetMapping("/configs") List<?> configs(){Set<String> available=Set.of("inspect_log","analyze_500_ips","validate_top_offending_ip","inspect_sql_schema","query_chicago_q1_revenue","validate_chicago_q1_revenue","inspect_employee_audit_schema","audit_employee_phase_1","calculate_new_york_median_salary","generate_employee_phase_3_report","validate_employee_audit");return configs.list().stream().filter(x->{for(JsonNode t:x.config().path("tools"))if(!available.contains(t.path("name").asText()))return false;return true;}).map(x->{Map<String,Object> m=new LinkedHashMap<>();m.put("config_ref",x.ref());m.put("sha256",x.sha256());m.put("agent_name",x.config().path("name").asText());m.put("tools",x.config().path("tools"));List<String> phases=new ArrayList<>();for(JsonNode p:x.config().at("/workflow/phases"))phases.add(p.path("id").asText());m.put("workflow_phases",phases);m.put("verbose_setup",x.config().at("/output/verbose_setup").asBoolean(false));return m;}).toList();}
 @PostMapping("/sessions") Mono<ResponseEntity<?>> create(@RequestBody CreateSessionRequest q){try{String goal=q.user_prompt()==null?"":q.user_prompt();if(goal.isEmpty())return Mono.just(ResponseEntity.unprocessableEntity().build());var c=configs.resolve(q.config_ref()==null?"test-case-1.yaml":q.config_ref());Set<String> available=Set.of("inspect_log","analyze_500_ips","validate_top_offending_ip","inspect_sql_schema","query_chicago_q1_revenue","validate_chicago_q1_revenue","inspect_employee_audit_schema","audit_employee_phase_1","calculate_new_york_median_salary","generate_employee_phase_3_report","validate_employee_audit");TreeSet<String> unknown=new TreeSet<>();for(JsonNode t:c.config().path("tools"))if(!available.contains(t.path("name").asText()))unknown.add(t.path("name").asText());if(!unknown.isEmpty())throw new IllegalArgumentException("config references unavailable compiled-in tools: "+String.join(", ",unknown));ObjectNode s=json.createObjectNode();s.put("id",UUID.randomUUID().toString());s.put("goal",goal);s.put("config_ref",c.ref());s.put("config_sha256",c.sha256());s.put("status","active");s.put("step_count",0);s.put("run_count",0);s.withArray("memory").addObject().put("role","user").put("content",goal);s.putArray("compaction_events");if(c.config().has("workflow")){ObjectNode w=s.putObject("workflow");w.put("active_phase",c.config().at("/workflow/entry_phase").asText());w.putArray("completed_phases");w.putArray("observed_evidence");w.putArray("transitions");}ObjectNode out=json.createObjectNode();out.set("session",s);out.put("message","Session created with immutable config "+c.ref()+"@"+c.sha256().substring(0,12)+".");return sessions.save(s,c.config(),event("session_created","config_ref",c.ref(),"config_sha256",c.sha256())).thenReturn(ResponseEntity.status(201).body(out));}catch(Exception e){return Mono.error(new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY,"Invalid config_ref: "+e.getMessage(),e));}}
 @GetMapping("/sessions/{session_id}") Mono<ResponseEntity<ObjectNode>> get(@PathVariable("session_id") String sessionId){return sessions.get(sessionId).map(x->ResponseEntity.ok(x.state())).defaultIfEmpty(ResponseEntity.notFound().build());}
 @PostMapping("/sessions/{session_id}/messages") Mono<ResponseEntity<?>> message(@PathVariable("session_id") String sessionId,@RequestBody AppendUserMessageRequest q){return sessions.get(sessionId).flatMap(x->{if(!"active".equals(x.state().path("status").asText()))return Mono.just(ResponseEntity.status(409).build());String content=q.content()==null?"":q.content();if(content.isEmpty())return Mono.just(ResponseEntity.unprocessableEntity().build());x.state().withArray("memory").addObject().put("role","user").put("content",content);ObjectNode o=json.createObjectNode();o.set("session",x.state());o.put("message","User input appended to session memory.");return sessions.save(x.state(),x.config(),event("user_message_appended")).thenReturn(ResponseEntity.ok(o));}).defaultIfEmpty(ResponseEntity.notFound().build());}
 @PostMapping("/sessions/{session_id}/run") Mono<ResponseEntity<ObjectNode>> run(@PathVariable("session_id") String sessionId,@RequestBody RunRequest q){
  return sessions.get(sessionId).flatMap(persisted -> {
   if(!"active".equals(persisted.state().path("status").asText())) return Mono.just(ResponseEntity.ok(response(persisted.state(),"Session is terminal; use POST /sessions/{session_id}/resume to continue it.")));
   return engine.run(persisted.state(),persisted.config(),steps(q.max_steps(),persisted.config()),e->sessions.save(persisted.state(),persisted.config(),e))
    .flatMap(done -> sessions.save(done,persisted.config(),event("run_finished")).thenReturn(ResponseEntity.ok(response(done,done.path("final_answer").asText("Run stopped.")))));
  }).defaultIfEmpty(ResponseEntity.notFound().build());
 }
 @PostMapping(value="/sessions/{session_id}/run/stream",produces=MediaType.TEXT_EVENT_STREAM_VALUE) Flux<ServerSentEvent<ObjectNode>> stream(@PathVariable("session_id") String sessionId,@RequestBody RunRequest q){return lifecycle(sessionId,q.max_steps(),null,false);}
 @PostMapping(value="/sessions/{session_id}/resume",produces=MediaType.TEXT_EVENT_STREAM_VALUE) Flux<ServerSentEvent<ObjectNode>> resume(@PathVariable("session_id") String sessionId,@RequestBody ResumeSessionRequest q){return lifecycle(sessionId,q.max_steps(),q.content(),true);}
 private Flux<ServerSentEvent<ObjectNode>> lifecycle(String sessionId,Integer maxSteps,String content,boolean resume){
  return Flux.create(sink -> sessions.get(sessionId).switchIfEmpty(Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND,"Session not found"))).flatMap(persisted -> {
   ObjectNode s=persisted.state();
   if(resume){s.put("status","active");s.remove("final_answer");if(content!=null&&!content.isEmpty())s.withArray("memory").addObject().put("role","user").put("content",content);}
   if(!resume&&!("active".equals(s.path("status").asText())))return Mono.error(new ResponseStatusException(HttpStatus.CONFLICT,"Session is terminal; use POST /sessions/{session_id}/resume to continue it"));
   Mono<Void> before=resume?sessions.save(s,persisted.config(),event("session_resumed","has_new_user_message",content!=null&&!content.isEmpty())).then(Mono.fromRunnable(()->sink.next(ServerSentEvent.builder(event("status","message","Session restored from PostgreSQL; resuming agent work.")).build()))):Mono.empty();
   ObjectNode setup=event("status","message",resume?"Restored immutable config, workflow evidence, memory, and compaction ledger.":"Agent initialized from persisted session state.");
   if(persisted.config().at("/output/verbose_setup").asBoolean(false))before=before.then(sessions.save(s,persisted.config(),setup).doOnSuccess(v->sink.next(ServerSentEvent.builder(setup).build())));
   return before.then(engine.run(s,persisted.config(),steps(maxSteps,persisted.config()),e->sessions.save(s,persisted.config(),e).doOnSuccess(v->sink.next(ServerSentEvent.builder(e).build()))))
    .flatMap(done->{ObjectNode finish=event("run_finished","status",done.path("status").asText());return sessions.save(done,persisted.config(),finish).doOnSuccess(v->{sink.next(ServerSentEvent.builder(finish).build());sink.next(ServerSentEvent.builder(event("status","message","Agent run finished.","status",done.path("status").asText())).build());});});
  }).subscribe(v->{},sink::error,sink::complete),FluxSink.OverflowStrategy.BUFFER);
 }
 private <T> Mono<T> within(String id,java.util.function.Supplier<Mono<T>> work){Object lock=locks.computeIfAbsent(id,k->new Object());return Mono.defer(()->{synchronized(lock){return work.get();}});}
 private int steps(Integer requested,ObjectNode c){if(requested!=null){int n=requested;if(n<1||n>50)throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY,"max_steps must be between 1 and 50");return n;}return c.at("/memory/max_steps").asInt();}private ObjectNode response(ObjectNode s,String m){ObjectNode o=json.createObjectNode();o.set("session",s);o.put("message",m);return o;}private ObjectNode event(String type,Object...kv){ObjectNode e=json.createObjectNode();e.put("type",type);for(int i=0;i<kv.length;i+=2)e.set(kv[i].toString(),json.valueToTree(kv[i+1]));return e;}
}