# agent-rust/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "State-Driven AI Agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/messages": {
      "post": {
        "summary": "Append User Message",
        "operationId": "append_user_message_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendUserMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/run": {
      "post": {
        "summary": "Run Session",
        "operationId": "run_session_sessions__session_id__run_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/run/stream": {
      "post": {
        "summary": "Stream Run Session",
        "operationId": "stream_run_session_sessions__session_id__run_stream_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/resume": {
      "post": {
        "summary": "Resume Session",
        "description": "Restore a terminal session, compact if needed, and stream its continuation.",
        "operationId": "resume_session_sessions__session_id__resume_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppendUserMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "Additional user input appended to the active session memory."
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "AppendUserMessageRequest"
      },
      "CreateSessionRequest": {
        "properties": {
          "user_prompt": {
            "type": "string",
            "minLength": 1,
            "title": "User Prompt",
            "description": "The user's goal; becomes the first session-memory message."
          },
          "config_ref": {
            "type": "string",
            "minLength": 1,
            "title": "Config Ref",
            "description": "Relative YAML path inside the server-approved config repository.",
            "default": "test-case-1.yaml"
          }
        },
        "type": "object",
        "required": [
          "user_prompt"
        ],
        "title": "CreateSessionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "system",
              "user",
              "assistant",
              "tool"
            ],
            "title": "Role"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "Message"
      },
      "ResumeSessionRequest": {
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "title": "ResumeSessionRequest",
        "description": "Continue a terminal session, optionally with a new user turn."
      },
      "RunRequest": {
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          }
        },
        "type": "object",
        "title": "RunRequest"
      },
      "SessionResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionState"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session",
          "message"
        ],
        "title": "SessionResponse"
      },
      "SessionState": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "config_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Ref"
          },
          "config_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Sha256"
          },
          "memory": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Memory"
          },
          "compacted_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compacted Context"
          },
          "compaction_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Compaction Events"
          },
          "workflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowState"
              },
              {
                "type": "null"
              }
            ]
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count",
            "default": 0
          },
          "run_count": {
            "type": "integer",
            "title": "Run Count",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "complete",
              "failed"
            ],
            "title": "Status",
            "default": "active"
          },
          "final_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Answer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal"
        ],
        "title": "SessionState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowState": {
        "properties": {
          "active_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Phase"
          },
          "completed_phases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Completed Phases"
          },
          "observed_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observed Evidence"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/WorkflowTransition"
            },
            "type": "array",
            "title": "Transitions"
          }
        },
        "type": "object",
        "required": [
          "active_phase"
        ],
        "title": "WorkflowState",
        "description": "Persistent workflow facts kept outside compactable raw message history."
      },
      "WorkflowTransition": {
        "properties": {
          "from_phase": {
            "type": "string",
            "title": "From Phase"
          },
          "to_phase": {
            "type": "string",
            "title": "To Phase"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "from_phase",
          "to_phase",
          "reason"
        ],
        "title": "WorkflowTransition"
      }
    }
  }
}